Bcast ci matrix elemenent-wise
MPI broadcast fails if the object is large. For cases with many determinants (> 16383 with OpenMPI 4.1.2 on Ubuntu), the `ci[0]` matrix can be large, which leads to failure in the `mpi::broadcast(world, ci, 0);` operation. This commit attempts to fix the issue by broadcasting the `ci[0]` matrix element-wise using a for loop.